move stubcontent and stuberror to their own class
