+++ YetAnotherHTTPStub/StubContent.swift
@@ -1,0 +1,29 @@
+//
+//  StubContent.swift
+//  YetAnotherHTTPStub
+//
+//  Created by Darren Lai on 7/23/17.
+//  Copyright © 2017 KinWahLai. All rights reserved.
+//
+
+import Foundation
+
+public enum StubContent: ExpressibleByNilLiteral, Equatable  {
+    public init(nilLiteral: ()) {
+        self = .noContent
+    }
+    
+    case data(Data)
+    case noContent
+}
+
+public func ==(lhs:StubContent, rhs:StubContent) -> Bool {
+    switch(lhs, rhs) {
+    case let (.data(lhsData), .data(rhsData)):
+        return (lhsData == rhsData) && lhsData.count == rhsData.count
+    case (.noContent, .noContent):
+        return true
+    default:
+        return false
+    }
+}
